Leitura de Qrcode e ajustes no adicionar item
leitura de Qrcode com MLKIT e o fluxo de editar, remover item.

+++ app/src/main/java/com/tech/building/domain/model/ItemRequestModel.kt
@@ -6,6 +6,6 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class ItemRequestModel(
     var materialModel: MaterialModel,
-    var qtdRequested: Double,
+    var qtdRequested: Int,
     var unit: String
 ) : Parcelable
